+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix Filtered drivers and clients in the history [#120](https://github.com/DevAlexandreCR/gorda-admin/pull/120)
+
 # Release Notes for 1.3.x
 
 ## [1.3.0(2024-02-03)](https://github.com/DevAlexandreCR/admin-driver/compare/1.3.0...1.2.10)