+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Now admin can enable and disable chatBot. [#114](https://github.com/DevAlexandreCR/admin-driver/pull/114)
+
 # Release Notes for 1.3.x
 
 ## [1.3.0(2024-02-03)](https://github.com/DevAlexandreCR/admin-driver/compare/1.3.0...1.2.10)