+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix when create a WpClient with the phone number with spaces [#122](https://github.com/DevAlexandreCR/gorda-admin/pull/122)
+
 # Release Notes for 1.3.x
 
 ## [1.3.0(2024-02-03)](https://github.com/DevAlexandreCR/admin-driver/compare/1.3.0...1.2.10)