+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
-
+- Fix when create a WpClient with the phone number with spaces [#122](https://github.com/DevAlexandreCR/gorda-admin/pull/122)
 - Fix Filtered drivers and clients in the history [#120](https://github.com/DevAlexandreCR/gorda-admin/pull/120)
 
 # Release Notes for 1.3.x